Steal MSL stuff from sms and fdlibm

+++ include/MSL_C/rand.h
@@ -1,12 +1,11 @@
 #ifndef RAND_H
 #define RAND_H
-#include "types.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void srand(u32 seed);
+void srand(unsigned long seed);
 int rand(void);
 
 #ifdef __cplusplus
